webhook_publisher/requirements.txt
argh==0.26.2
asn1crypto==0.24.0
attrs==17.4.0
authzerolib>=1.0.3
aws-encryption-sdk==1.3.4
aws-xray-sdk==0.95
boto==2.48.0
boto3==1.7.5
botocore==1.10.5
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
configobj==5.0.6
cookies==2.2.1
cryptography==2.2.2
docker==3.2.1
docker-pycreds==0.2.2
docutils==0.14
ecdsa==0.13
everett==0.9
flake8==3.5.0
Flask==0.12.2
Flask-Cors==3.0.3
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsondiff==1.1.1
jsonpickle==0.9.6
jsonschema==2.6.0
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
moto==1.3.3
# git+git://github.com/mozilla-iam/cis.git@v0.1.2.1#egg=mozilla-iam-cis
nose-watch==0.9.2
pathtools==0.1.2
pbr==4.0.2
pluginbase==0.5
pyaml==17.12.1
pyblake2==1.1.2
pycodestyle==2.3.1
pycparser==2.18
pycryptodome==3.3.1
pyflakes==1.6.0
pykmssig==0.0.8
python-dateutil==2.6.1
python-decouple==3.1
python-dotenv==0.8.2
python-jose-cryptodome==1.3.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
responses==0.9.0
s3transfer==0.1.13
six==1.11.0
urllib3==1.22
watchdog==0.8.3
watchtower==0.5.3
websocket-client==0.47.0
Werkzeug==0.14.1
wrapt==1.10.11
xmltodict==0.11.0
